+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 ### Features
 * change `append!` from macro to function (it will work with any value as fist list argument not only with names
 * `help` function and help system inside functions. First expression after function arguments is docs.
-* add `pipe`, `compose` and `some` function
+* add `pipe`, `compose`, `fold` and `some` functions
 * new function `unbind` to help write function that work with functions as objects (LIPS wrap them)
 * add alias for `.` as `get` so you can pass it to curry, with `.` LIPS will try to create pair
 * lambda function introspection
@@ -14,6 +14,7 @@
 * better loading of code
 * make `map`, `filter`, `find`, `for-each` and `reduce` sync when possible
 * basic type checking
+* documented `unquote-splicing` and `unquote` functions that throw error when used outside of `quasiquote`
 ### Bug fixes
 * fix `if` it should not return false if false value is undefined
 * fix even and odd functions with BigInt